Add files via upload
More detailed instructions about how to install and use this site skin will be in its Ao3 work and tumblr post. Basic instructions are also included in the respective codes for the skin and its mobile fixes. Note that any of those included notes will be stripped by Ao3's formatting. The fonts must be downloaded and installed to view this skin entirely as pictured and intended, but it is usable without.